move train scripts around
